requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_tflow_select=2.2.0
abseil-cpp=20211102.0
absl-py=1.3.0
aiohttp=3.8.3
aiosignal=1.2.0
appdirs=1.4.4
appnope=0.1.3
asttokens=2.2.1
astunparse=1.6.3
async-timeout=4.0.2
attrs=22.1.0
backcall=0.2.0
backports=1.0
backports.functools_lru_cache=1.6.4
blas=1.0
blinker=1.4
bottleneck=1.3.5
breadability=0.1.20
brotli=1.0.9
brotli-bin=1.0.9
brotlipy=0.7.0
c-ares=1.18.1
ca-certificates=2023.01.10
cachetools=4.2.2
catalogue=2.0.7
certifi=2022.12.7
cffi=1.15.1
chardet=5.1.0
charset-normalizer=2.0.4
click=8.0.4
colorama=0.4.6
comm=0.1.2
contourpy=1.0.5
cryptography=38.0.4
cycler=0.11.0
cymem=2.0.6
cython-blis=0.7.7
debugpy=1.6.6
decorator=5.1.1
docopt=0.6.2
en-core-web-sm=3.3.0
et_xmlfile=1.1.0
executing=1.2.0
fftw=3.3.9
filelock=3.9.0
flatbuffers=2.0.0
fonttools=4.25.0
freetype=2.12.1
frozenlist=1.3.3
fst-pso=1.8.1
funcy=2.0
future=0.18.2
fuzzytm=2.0.5
gast=0.4.0
gensim=4.3.0
giflib=5.2.1
google-auth=2.6.0
google-auth-oauthlib=0.4.4
google-pasta=0.2.0
grpc-cpp=1.46.1
grpcio=1.42.0
h5py=3.7.0
hdf5=1.10.6
huggingface_hub=0.10.1
icu=58.2
idna=3.4
importlib-metadata=6.0.0
importlib_metadata=6.0.0
iniconfig=1.1.1
intel-openmp=2021.4.0
ipykernel=6.21.2
ipython=8.10.0
jedi=0.18.2
jinja2=3.1.2
joblib=1.2.0
jpeg=9e
jupyter_client=8.0.3
jupyter_core=5.2.0
keras=2.10.0
keras-preprocessing=1.1.2
kiwisolver=1.4.4
krb5=1.19.4
langcodes=3.3.0
lcms2=2.12
lerc=3.0
libbrotlicommon=1.0.9
libbrotlidec=1.0.9
libbrotlienc=1.0.9
libcurl=7.87.0
libcxx=14.0.6
libdeflate=1.8
libedit=3.1.20221030
libev=4.33
libffi=3.4.2
libgfortran=5.0.0
libgfortran5=11.3.0
libiconv=1.16
libnghttp2=1.46.0
libopenblas=0.3.21
libpng=1.6.37
libprotobuf=3.20.3
libsodium=1.0.18
libssh2=1.10.0
libtiff=4.5.0
libuv=1.44.2
libwebp=1.2.4
libwebp-base=1.2.4
libxml2=2.10.3
libxslt=1.1.37
llvm-openmp=14.0.6
lxml=4.9.2
lz4-c=1.9.4
markdown=3.4.1
markupsafe=2.1.1
matplotlib=3.6.2
matplotlib-base=3.6.2
matplotlib-inline=0.1.6
miniful=0.0.6
mkl=2021.4.0
mkl-service=2.4.0
mkl_fft=1.3.1
mkl_random=1.2.2
multidict=6.0.2
munkres=1.1.4
murmurhash=1.0.7
ncurses=6.4
nest-asyncio=1.5.6
ninja=1.10.2
ninja-base=1.10.2
nltk=3.7
numexpr=2.8.4
numpy=1.23.5
numpy-base=1.23.5
oauthlib=3.2.1
openpyxl=3.0.10
openssl=1.1.1t
opt_einsum=3.3.0
ordered-set=4.1.0
packaging=22.0
pandas=1.5.2
parso=0.8.3
pathy=0.10.1
patsy=0.5.3
pexpect=4.8.0
pickleshare=0.7.5
pillow=9.3.0
pip=22.3.1
platformdirs=3.0.0
pluggy=1.0.0
pooch=1.4.0
preshed=3.0.6
prompt-toolkit=3.0.36
protobuf=3.20.3
psutil=5.9.4
ptyprocess=0.7.0
pure_eval=0.2.2
py=1.11.0
pyasn1=0.4.8
pyasn1-modules=0.2.8
pycountry=22.3.5
pycparser=2.21
pydantic=1.8.2
pyfume=0.2.25
pygments=2.14.0
pyjwt=2.4.0
pylatex=1.4.1
pyldavis=3.4.0
pyopenssl=22.0.0
pyparsing=3.0.9
pysocks=1.7.1
pytest=7.1.2
python=3.9.16
python-dateutil=2.8.2
python-dotenv=1.0.0
python-flatbuffers=2.0
python_abi=3.9
pytorch=1.12.1
pytz=2022.7
pyyaml=6.0
pyzmq=25.0.0
re2=2022.04.01
readline=8.2
regex=2022.7.9
requests=2.28.1
requests-oauthlib=1.3.0
rsa=4.7.2
scikit-learn=1.2.1
scipy=1.10.0
seaborn=0.12.2
setuptools=65.6.3
shellingham=1.5.0
simpful=2.10.0
six=1.16.0
smart_open=5.2.1
snappy=1.1.9
spacy=3.3.1
spacy-legacy=3.0.9
spacy-loggers=1.0.1
sqlite=3.40.1
srsly=2.4.6
stack_data=0.6.2
statsmodels=0.13.5
sumy=0.11.0
tensorboard=2.10.0
tensorboard-data-server=0.6.1
tensorboard-plugin-wit=1.6.0
termcolor=2.1.0
textblob=0.15.3
thinc=8.0.15
threadpoolctl=2.2.0
tk=8.6.12
tokenizers=0.11.4
tomli=2.0.1
torchvision=0.13.1
tornado=6.2
tqdm=4.64.1
traitlets=5.7.1
transformers=4.24.0
typer=0.4.1
typing-extensions=4.4.0
typing_extensions=4.4.0
tzdata=2022g
urllib3=1.26.14
wasabi=0.9.1
wcwidth=0.2.6
werkzeug=2.2.2
wheel=0.35.1
wordcloud=1.8.2.2
wrapt=1.14.1
xz=5.2.10
yaml=0.2.5
yarl=1.8.1
zeromq=4.3.4
zipp=3.14.0
zlib=1.2.13
zstd=1.5.2
